Fix bug and change start screen page view

+++ src/main/java/ir/comprehensive/service/WarehouseService.java
@@ -4,11 +4,15 @@ import ir.comprehensive.domain.Warehouse;
 import ir.comprehensive.domain.WarehouseTag;
 import ir.comprehensive.model.WarehouseModel;
 import ir.comprehensive.model.WarehouseTagModel;
+import ir.comprehensive.repository.ProductDeliveryRepository;
 import ir.comprehensive.repository.WarehouseRepository;
+import ir.comprehensive.repository.WarehouseTagRepository;
 import ir.comprehensive.service.extra.GeneralException;
 import ir.comprehensive.service.extra.Swappable;
 import ir.comprehensive.utils.MessageUtils;
 import ir.comprehensive.utils.StringUtils;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,9 +28,13 @@ import java.util.stream.Collectors;
 @Transactional
 public class WarehouseService implements Swappable<Warehouse> {
     private WarehouseRepository repository;
+    private ProductDeliveryRepository productDeliveryRepository;
+    private WarehouseTagRepository warehouseTagRepository;
 
-    public WarehouseService(WarehouseRepository repository) {
+    public WarehouseService(WarehouseRepository repository, ProductDeliveryRepository productDeliveryRepository, WarehouseTagRepository warehouseTagRepository) {
         this.repository = repository;
+        this.productDeliveryRepository = productDeliveryRepository;
+        this.warehouseTagRepository = warehouseTagRepository;
     }
 
     public Optional<Warehouse> load(Long id) throws GeneralException {
@@ -35,6 +43,11 @@ public class WarehouseService implements Swappable<Warehouse> {
             throw new GeneralException("null id");
         }
         return repository.findById(id);
+    }
+
+    public Optional<List<Warehouse>> findByName(String name) {
+        Page<Warehouse> warehouses = repository.findByName(name, PageRequest.of(0, 10));
+        return Optional.of(warehouses.getContent());
     }
 
     public Optional<List<Warehouse>> loadAll() {
@@ -54,6 +67,9 @@ public class WarehouseService implements Swappable<Warehouse> {
             if (searchExample.getCompanyName() != null && !searchExample.getCompanyName().isEmpty()) {
                 predicateList.add(criteriaBuilder.like(root.get("companyName"), StringUtils.makeAnyMatch(searchExample.getCompanyName())));
             }
+            if (searchExample.getCategory() != null && searchExample.getCategory().getId() != null) {
+                predicateList.add(criteriaBuilder.equal(root.get("category").get("id"), searchExample.getCategory().getId()));
+            }
             if (searchExample.getTagList() != null && !searchExample.getTagList().isEmpty()) {
                 Join<Warehouse, WarehouseTag> categories = root.join("tagList");
                 predicateList.add(categories.in(searchExample.getTagList().stream().map(WarehouseTagModel::getId).collect(Collectors.toSet())));
@@ -61,7 +77,7 @@ public class WarehouseService implements Swappable<Warehouse> {
             return criteriaBuilder.and(predicateList.toArray(new Predicate[predicateList.size()]));
         };
 
-        return Optional.of(repository.findAll(warehouseSpecification));
+        return Optional.of(repository.findAll(warehouseSpecification).stream().distinct().collect(Collectors.toList()));
     }
 
     private void validateEntity(Warehouse warehouse) throws GeneralException {
@@ -91,6 +107,20 @@ public class WarehouseService implements Swappable<Warehouse> {
     }
 
     public Optional<Warehouse> saveOrUpdate(Warehouse warehouse) throws GeneralException {
+        List<WarehouseTag> newTagListForSave = new ArrayList<>();
+
+        if (warehouse.getTagList() != null) {
+            for (WarehouseTag warehouseTag : warehouse.getTagList()) {
+                if (warehouseTag.getId() != null) {
+                    newTagListForSave.add(warehouseTagRepository.findById(warehouseTag.getId()).orElse(null));
+                } else {
+                    newTagListForSave.add(warehouseTag);
+                }
+            }
+        }
+
+        warehouse.setTagList(newTagListForSave);
+
         return warehouse.getId() == null ? save(warehouse) : update(warehouse);
     }
 
@@ -100,8 +130,12 @@ public class WarehouseService implements Swappable<Warehouse> {
             // TODO fix message
             throw new GeneralException("not null id");
         }
+        if (productDeliveryRepository.isWarehouseExist(id)) {
+            throw new GeneralException(MessageUtils.Message.PRODUCT + " " + MessageUtils.Message.USE_IN + " " + MessageUtils.Message.STOREROOM);
+        }
 
         repository.deleteById(id);
+
         return Optional.of(id);
     }
 }
